Moved examples into their own respective subfolder.
mesh.py is the file to run to generate the mesh.

model.py is executed to run the amigo model on the generated mesh.
